pyproject.toml
[project]
name = "fastapi-base"
version = "0.1.0"
description = "FastAPI template with SQLModel and authentication support"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "python-dotenv>=1.0.1",
    "requests>=2.32.5",
    "boto3>=1.42.2",
    "python-multipart>=0.0.20",
    "psycopg2-binary>=2.9.11",
    "alembic>=1.17.2",
    "websockets>=15.0.1",
    "python-jose[cryptography]>=3.5.0",
    "passlib>=1.7.4",
    "bcrypt>=5.0.0",
    "aiosmtplib>=5.0.0",
    "email-validator>=2.3.0",
    "jinja2>=3.1.6",
    "redis>=7.1.0",
    "prometheus-client>=0.23.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
]

[project.scripts]
start = "main:main"
